+++ test/client_test.rb
@@ -382,6 +382,35 @@ class ASCToolingClientTest < Minitest::Test
     assert_equal "1", captured[:params]["limit"]
   end
 
+  def test_find_latest_eligible_build_uses_server_side_platform_and_state_filters
+    client = ASCTooling::Client.allocate
+    captured = nil
+    client.define_singleton_method(:request_json) do |method, path, params: nil, body: nil|
+      captured = { method: method, path: path, params: params, body: body }
+      {
+        "data" => [
+          {
+            "id" => "build-eligible",
+            "attributes" => { "version" => "2026082501" }
+          }
+        ]
+      }
+    end
+
+    build = client.find_latest_eligible_build("app-123", "1.3.0", platform: "IOS")
+
+    assert_equal "build-eligible", build["id"]
+    assert_equal "GET", captured[:method]
+    assert_equal "/v1/builds", captured[:path]
+    assert_equal "app-123", captured[:params]["filter[app]"]
+    assert_equal "1.3.0", captured[:params]["filter[preReleaseVersion.version]"]
+    assert_equal "IOS", captured[:params]["filter[preReleaseVersion.platform]"]
+    assert_equal "VALID", captured[:params]["filter[processingState]"]
+    assert_equal "APP_STORE_ELIGIBLE", captured[:params]["filter[buildAudienceType]"]
+    assert_equal "-uploadedDate", captured[:params]["sort"]
+    assert_equal "1", captured[:params]["limit"]
+  end
+
   def test_paginated_resources_follows_next_link_until_exhausted
     client = ASCTooling::Client.allocate
     requests = []